fix use organization_endpoints for /organization management
